fix: separate columns and components, useMemo (#68)
