+++ src/components/DataItem/SectorBadge.tsx
@@ -2,7 +2,6 @@ import { Sectors } from '@/defaults'
 import { Badge } from '@ttab/elephant-ui'
 import { cn } from '@ttab/elephant-ui/utils'
 
-
 export const SectorBadge = ({ color, value }: { color?: string, value: string }): JSX.Element => {
   const sector = Sectors.find((label) => label.value === value)
   const chosenColor = color || sector?.color
