+++ src/views/PlanningOverview/PlanningTable/Columns/Assignees.tsx
@@ -1,19 +1,6 @@
-import { type ColumnDef } from '@tanstack/react-table'
-import { type Planning } from '../data/schema'
 import { AssigneeAvatars } from '@/components/DataItem/AssigneeAvatars'
-import { Users } from '@ttab/elephant-ui/icons'
+import { useMemo } from 'react'
 
-export const assignees: ColumnDef<Planning> = {
-  id: 'assignees',
-  meta: {
-    filter: 'facet',
-    name: 'Assignees',
-    columnIcon: Users,
-    className: 'box-content w-[112px] hidden @5xl/view:[display:revert]'
-  },
-  accessorFn: (data) => data._source['document.meta.core_assignment.rel.assignee.name'],
-  cell: ({ row }) => {
-    const assignees = row.getValue<string[]>('assignees') || []
-    return <AssigneeAvatars assignees={assignees} />
-  }
+export const Assignees = ({ assignees }: { assignees: string[] }): JSX.Element => {
+  return useMemo(() => <AssigneeAvatars assignees={assignees} />, [assignees])
 }
