+++ src/views/PlanningOverview/PlanningTable/Columns/Time.tsx
@@ -1,19 +1,6 @@
-import { type ColumnDef } from '@tanstack/react-table'
-import { type Planning } from '../data/schema'
-import { TimeDisplay } from '../../../../components/DataItem/TimeDisplay'
-import { getPublishTime } from '@/lib/getPublishTime'
-import { Clock } from '@ttab/elephant-ui/icons'
+import { TimeDisplay } from '@/components/DataItem/TimeDisplay'
+import { useMemo } from 'react'
 
-export const time: ColumnDef<Planning> = {
-  id: 'time',
-  meta: {
-    filter: null,
-    name: 'Time',
-    columnIcon: Clock,
-    className: 'box-content w-[120px] hidden @3xl/view:[display:revert]'
-  },
-  accessorFn: (data) => getPublishTime(data._source['document.meta.core_assignment.data.publish']),
-  cell: ({ row }) => (
-    <TimeDisplay date={row.getValue('time')} />
-  )
+export const Time = ({ date }: { date: Date }): JSX.Element => {
+  return useMemo(() => <TimeDisplay date={date} />, [date])
 }
