+++ src/views/PlanningOverview/PlanningTable/index.tsx
@@ -1,5 +1,4 @@
 import React, { useEffect } from 'react'
-import { columns } from '@/views/PlanningOverview/PlanningTable/Columns'
 import {
   type ColumnDef,
   flexRender
@@ -10,6 +9,7 @@ import { Toolbar } from './Toolbar'
 import { useView } from '@/hooks'
 import { isEditableTarget } from '@/lib/isEditableTarget'
 import { useTable } from '@/hooks/useTable'
+import { columns } from './Columns'
 
 interface PlanningTableProps<TData, TValue> {
   columns: Array<ColumnDef<TData, TValue>>
@@ -123,7 +123,7 @@ export const PlanningTable = <TData, TValue>({
     <div className="space-y-2">
       <Toolbar table={table} />
       <div className="rounded-md">
-        <Table>
+        <Table className='table-fixed'>
           <TableBody>
             <TableBodyElement />
           </TableBody>
